Don't send numeric keypad events to both keyboard and special callbacks.
When NumLock is down, the XK_KP_Left family of keys would both get
translated to regular text "4" sent to the keyboard callback, and also
passed raw to the special callback.

Change the X11 code to never send the special key event if we sent the
translated unicode equivalent to the keyboard callback.